% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_inspire_grids.R
\name{calculate_inspire_grids}
\alias{calculate_inspire_grids}
\title{Calculate 1km and 100m INSPIRE Grid IDs from coordinates and add to data.frame}
\usage{
calculate_inspire_grids(coords.file = "", coords.object = "")
}
\arguments{
\item{coords.file}{Path to file with coordinates; must be formatted as
csv text file with a 'X' and header...}

\item{coords.object}{An R object ("Rdata" or "rda") containing the
coordinates}
}
\value{
data.frame
}
\description{
Calculate 1km and 100m INSPIRE Grid IDs from coordinates and add to data.frame
}
