% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_check_crs.R
\name{cdr_check_crs}
\alias{cdr_check_crs}
\title{Check for missing coordinate systems in shape files and export it to text
file}
\usage{
cdr_check_crs(data.path, out.file)
}
\arguments{
\item{data.path}{Directory to check for missing CRS}

\item{out.file}{Location and name of output text file}
}
\description{
Check for missing coordinate systems in shape files and export 
it to text file. Initially developed for environmental noise data from the 
CDR it also can be used for all locations storing a bunch of ESRI shapefiles.
}
