% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_download_single.R
\name{cdr_download_single}
\alias{cdr_download_single}
\title{Download single layers of environmental noise data for individual German
federal states in the CDR}
\usage{
cdr_download_single(data.path = ".",
  url.path = c(paste("https://cdr.eionet.europa.", "eu/de/eu/noise/df8/",
  "colvjswog/envvlltfa/", sep = ""), paste("https://cdr.eionet.europa.",
  "eu/de/eu/noise/df8/", "colvjswog/envvk89vq/", sep = "")),
  file = c("DE_NW_ag14_Aggroad_Lden", "DE_BY_ag1_Aggroad_Lden"))
}
\arguments{
\item{data.path}{Place where the data will be stored}

\item{url.path}{Vector of urls that point to data subfolders of German
federal states in the CDR}

\item{file}{Vector of filenames (\code{length == url.path}) that should be
downloaded}
}
\description{
Download single layers of environmental noise data for individual German
federal states in the CDR
}

