% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_harmonize.R
\name{cdr_harmonize}
\alias{cdr_harmonize}
\title{Harmonize CDR data}
\usage{
cdr_harmonize(data.path = ".", which = "road", when = "lden",
  crs = "+init=epsg:3035")
}
\arguments{
\item{data.path}{Folder where the CDR shapefiles live}

\item{which}{Which noise type should be harmonized}

\item{when}{Which daytime of noise should be harmonized}

\item{crs}{Definition of coordinate reference system}
}
\description{
Harmonize CDR data
}
