% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_linking_simple.R
\name{cdr_linking_simple}
\alias{cdr_linking_simple}
\title{Link Coordinates to data downloaded from the CDR}
\usage{
cdr_linking_simple(file = "", coords.file = "")
}
\arguments{
\item{coords.file}{Path to file with coordinates; must be formatted as
csv text file with a 'X' andheader...}
}
\value{
A \code{data.frame} with environmental noise attributes for each 
        coordinate
}
\description{
This function performs spatial linking based on coordinates data
and environmental noise data from the CDR. By default it uses
simulated coordinates data for demonstration purposes. Own coordinates, however,
can be used as well.
}
