% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_download.R
\name{census_download_1km}
\alias{census_download_1km}
\title{Download}
\usage{
census_download_1km(data.path = ".")
}
\arguments{
\item{data.path}{Place where the data will be stored}
}
\description{
Download
}
