% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_linking_100m.R
\name{census_linking_100m}
\alias{census_linking_100m}
\title{Link Coordinates to German Census 2011 Data}
\usage{
census_linking_100m(own.data = FALSE, set.missings = TRUE,
  data.path = ".", coords.file = "", coords.object = "", suffix = "")
}
\arguments{
\item{own.data}{Use own already downloaded census data (calls
\code{georefum::rasterize.census()}, original filenames must be preserved)}

\item{data.path}{Has to be defined when using option \code{download =} or
\code{own.data =}}

\item{coords.file}{Path to file with coordinates; must be formatted as
csv text file with a 'X' andheader...}

\item{coords.object}{An R object ("Rdata" or "rda") containing the coordinates}

\item{which}{Vector of Census attributeses that should be linked to the data}

\item{set.missing}{Toggle whether to set missings on Census attributes}
}
\value{
A \code{data.frame} with census attributes for each coordinate
}
\description{
This function performs spatial linking based on coordinates data
and raster grid cell data from the German Census 2011. By default it uses
simulated coordinates data for demonstration purposes. Own coordinates, however,
can be used as well.
}
