% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_linking_simple.R
\name{census_linking_simple}
\alias{census_linking_simple}
\title{Link Coordinates to German Census 2011 Data}
\usage{
census_linking_simple(download = FALSE, own.data = FALSE,
  which = c("Einwohner", "Alter_D", "unter18_A", "ab65_A", "Auslaender_A",
  "HHGroesse_D", "Leerstandsquote", "Wohnfl_Bew_D", "Wohnfl_Whg_D"),
  set.missings = TRUE, data.path = ".", coords.file = "",
  coords.object = "")
}
\arguments{
\item{download}{Download census data (calls
\code{georefum::download.census()})}

\item{own.data}{Use own already downloaded census data (calls
\code{georefum::rasterize.census()}, original filenames must be preserved)}

\item{which}{Vector of Census attributeses that should be linked to the data}

\item{data.path}{Has to be defined when using option \code{download =} or
\code{own.data =}}

\item{coords.file}{Path to file with coordinates; must be formatted as
csv text file with a 'X' andheader...}

\item{coords.object}{An R object ("Rdata" or "rda") containing the coordinates}

\item{set.missing}{Toggle whether to set missings on Census attributes}
}
\value{
A \code{data.frame} with census attributes for each coordinate
}
\description{
Link Coordinates to German Census 2011 Data
}
