% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_rasterize.R
\name{census_rasterize}
\alias{census_rasterize}
\title{Convert German Census 2011 files into raster grids files}
\usage{
census_rasterize(data.path = ".", download = FALSE, save = FALSE,
  save.format = "GTiff", save.overwrite = FALSE)
}
\arguments{
\item{data.path}{Has to be defined when using option \code{download =} or
\code{own.data =}}

\item{download}{Download census data (calls
\code{georefum::download.census()})}

\item{save}{Toggle whether Census rasters should be stored as individual
raster files on hard disk}

\item{save.format}{Define which raster data format should be used for
storing}

\item{save.overwrite}{Will overwrite existing files if they exist}
}
\description{
Convert German Census 2011 files into raster grids files
}

