% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_shapefiles_convert.R
\name{census_shapefiles_convert}
\alias{census_shapefiles_convert}
\title{convert shapefiles}
\usage{
census_shapefiles_convert(inspire.grids.convert = TRUE,
  municipalities.convert = TRUE, download = TRUE, data.path = ".")
}
\arguments{
\item{inspire.grids.convert}{Toggle conversion of inspire grids}

\item{municipalities.convert}{Toggle conversion of municipalties shapefile}

\item{download}{Toggle download of inspire grids and municipalties shapefile}

\item{data.path}{Place where downloaded files should be stored (only relevant
when \code{download} is toggled)}
}
\description{
convert shapefiles
}

