% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_key_database.R
\name{create_key_database}
\alias{create_key_database}
\title{Create keys}
\usage{
create_key_database(surv.dat = NULL, coord.dat = NULL)
}
\arguments{
\item{surv.dat}{Survey data Object}

\item{coord.dat}{Coordinates Data Object, stores as SpatialPointsDataFrame}
}
\description{
Create keys
}
