% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_keys.R
\name{create_keys}
\alias{create_keys}
\title{Create keys}
\usage{
create_keys(use.existing.surv.dat = "", N = 10000)
}
\arguments{
\item{use.existing.surv.dat}{Path to survey data that is stored in the comma
separated values (CSV) file format; a variable/attribute/column named "id"
with uniquie IDs ist required, otherwise the function will fail}

\item{N}{Number of unique IDs that should be computed}
}
\description{
Create keys
}
