% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census.R
\name{download_census_1km}
\alias{download_census_1km}
\title{Download}
\usage{
download_census_1km(data.path = ".")
}
\arguments{
\item{data.path}{Place where the data will be stored}
}
\description{
Download
}

