% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.dat.R
\docType{data}
\name{key.dat}
\alias{key.dat}
\title{Simulated Key Database}
\format{A data.frame with 10000 rows and 2 variables:
\describe{
  \item{id}{id for individual respondent}
  \item{id2}{second id for individual respondent that is used to merge
  with spatial attributes}
}}
\source{
\code{georefum::simulate_surveydata()}, then
\code{georefum::create_keys()}
}
\usage{
data(georefum::key.dat)
}
\description{
A dataset containing two id variables for 10000 individual respondents
}
\keyword{datasets}

