% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_coordinates.R
\name{simulate_coordinates}
\alias{simulate_coordinates}
\title{Simulate coordinates}
\usage{
simulate_coordinates(mun.N = 100, resp.N = 100, mun = "")
}
\arguments{
\item{mun.N}{Number of municipalties that should be sampled}

\item{resp.N}{Number of respondents that should be sampled within the already
sampled municipalities}
}
\description{
Simulate coordinates
}
