% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_surveydata.R
\name{simulate_surveydata}
\alias{simulate_surveydata}
\title{Simulate survey data}
\usage{
simulate_surveydata(N = 10000)
}
\arguments{
\item{N}{Number of respondents that should be simulated}
}
\description{
Simulate survey data
}

