% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_from_inspire.R
\name{xy_from_inspire}
\alias{xy_from_inspire}
\title{Retreive centroid geocoordiantes from 1km² INSPIRE Grid IDs from coordinates and add to data.frame}
\usage{
xy_from_inspire(id.file = "", id.object = "", col.name = "INSPIRE_1km")
}
\arguments{
\item{id.file}{Path to file with INSPIRE IDs as one column; must be formatted as
csv text file with a ';' as separator}

\item{id.object}{An R object ("Rdata" or "rda") containing the
INSPIRE IDs as one column; must be loaded into the environment}
}
\value{
data.frame
}
\description{
Retreive centroid geocoordiantes from 1km² INSPIRE Grid IDs from coordinates and add to data.frame
}
