% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_download_whole.R
\name{cdr_download_whole}
\alias{cdr_download_whole}
\title{Download all available layers of environmental noise data for all German
federal states in the CDR}
\usage{
cdr_download_whole(data.path = ".")
}
\arguments{
\item{data.path}{Place where the data will be stored}
}
\description{
Download all available layers of environmental noise data for all German
federal states in the CDR
}
