% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_linking_simple.R
\name{cdr_linking_simple}
\alias{cdr_linking_simple}
\title{Link Coordinates to data downloaded from the CDR}
\usage{
cdr_linking_simple(file = "", coords.file = "")
}
\arguments{
\item{coords.file}{Path to file with coordinates; must be formatted as
csv text file with a 'X' andheader...}
}
\value{
A \code{data.frame} with census attributes for each coordinate
}
\description{
Link Coordinates to data downloaded from the CDR
}

