% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_linking_focal.R
\name{census_linking_focal}
\alias{census_linking_focal}
\title{Link Coordinates to German Census 2011 Data by Using Focal Analyses}
\usage{
census_linking_focal(download = FALSE, own.data = FALSE,
  which = c("Einwohner", "Alter_D", "unter18_A", "ab65_A", "Auslaender_A",
  "HHGroesse_D", "Leerstandsquote", "Wohnfl_Bew_D", "Wohnfl_Whg_D"),
  set.missings = TRUE, data.path = ".", coords.file = "",
  coords.object = "", focal.matrix = matrix(c(1, 1, 1, 1, 1, 1, 1, 1, 1), nr
  = 3, nc = 3), fun = "mean")
}
\arguments{
\item{download}{Download census data (calls
\code{georefum::download.census()})}

\item{own.data}{Use own already downloaded census data (calls
\code{georefum::rasterize.census()}, original filenames must be preserved)}

\item{which}{Vector of Census attributeses that should be linked to the data}

\item{data.path}{Has to be defined when using option \code{download =} or
\code{own.data =}}

\item{coords.file}{Path to file with coordinates; must be formatted as
csv text file with a 'X' andheader...}

\item{coords.object}{An R object ("Rdata" or "rda") containing the coordinates}

\item{focal.matrix}{Matrix on which upon focal analyses are computed; cells
can be weighted as well as set to \code{NA} (for details see function
\code{raster::focal()})}

\item{fun}{Function that is used in the focal analyses}

\item{set.missing}{Toggle whether to set missings on Census attributes}
}
\value{
A \code{data.frame} with census attributes based on focal analyses
for each coordinate
}
\description{
Link Coordinates to German Census 2011 Data by Using Focal Analyses
}

